Year 2020 - Day 04
Originally tried to implement using a std.StringHashMap, but the code
seems to run against a zig stdlib bug; left the previous commit in place
in case I want to look into this bug later.

So we now use an enum for the field type and an array indexed on this
enum to summarize the counts, making the code much simpler and reusable.