Day 10: use dynamic programming
Previously we had found a formula based on the products of certain
triangular numbers given the counts of consecutive gaps of 1 in the
ratings of successive adapters.  The problem is that it was not clear at
all where this came from.

We now use traditional dynamic programming, to good results.