2022 day 19
This is an awful, ugly, randomized brute-force solution.

I basically could not be arsed to come up with a good pruning strategy
for selecting the robots to build, as well as getting rid as early as
possible of useless tree branches.  Instead, I just choose randomly
whether to build each type of robot and run it a gazillion times,
keeping track of the best results so far.

On my M1 laptop part 1 takes 19s and part 2 takes 30s.

Shameful.