2023 day 14: optimize
I usually use a HashMap to represent a grid, because it is not always
obvious how big the grid will have to be. But I decided to check how
much of a performance impact that has, vs using a matrix.  It is quite
significant: the time for part 2 went from 4s to 350ms (10x).