Read whole input file into memory
This simplifies a lot of the stuff.

I also got tired of writing code that sometimes needs to remember the
state (line) while parsing the input -- this is now gone because we get
the whole input to parse at once.

I also centralize in a single place (command.zig) some policies:
* Which allocator we use
* How we measure time
* How we print answers